Split the UpdateProcessor into separat interfaces
There are now two main interfaces:
 - EventDispatcher, which takes new RunnerEvents (formerly known as
   UpdateInformation) and dispatches them back to the main coordinator
 - SleepTimeProvider, which provides the Runner with the next time to
   sleep (currently just returning random numbers)

In the process of splitting those functionality I created a message
broker which will act as a new layer between the runner coordination
logic and the front end.

Basically, there are three layers in the application. The first, lowest
one, represents all runners. Each of them executes in a separate thread,
and can only communicate with the coordinator via limited interfaces.
The second layer is the coordinator, which manages the runners (what
round it currently is in, the time it took to complete a round, who is
the winner, and so on). The coordinator does not have a dedicated thread
but rather runs "shared" on all runner threads. It's functionality is
backed by a set of concurrent collections.
The top most layer can be seen as the front end, it runs on the main
thread (automatically created by the java platform) and receives updates
about the overall process through the message broker.

The above model is the goal, but currently (as of this commit) not fully
implemented.

Signed-off-by: Niklas Wimmer <nwimmer@student.tgm.ac.at>