Decouple the front end from the runner managing
In the process of doing this I already implemented something which was
planned for later: the runner no longer manages itself in which round it
is in.

To communicate between the manager and the front end a message
distributor gets used, which runs in its own thread and has as its only
job to poll the message broker and distribute all new messages to
everyone who is interested (observer pattern).

Signed-off-by: Niklas Wimmer <nwimmer@student.tgm.ac.at>